feat: add APIs to get one user and get all users. Edit start script. Fix request validator for update user API.

+++ src/service/UserService.js
@@ -8,19 +8,13 @@ const {
   Unauthorized,
   Forbidden,
   InternalServerError,
+  BadRequest,
 } = require('../middleware/errorHandler');
 const MailService = require('./MailingService');
 const { jwtSecret } = require('../../config');
 const { startSession } = require('mongoose');
 
-async function createUser(data) {
-  Object.assign(data, {
-    emailConfirmed: true,
-  });
-  const user = await User.create(data);
-  return user;
-}
-
+// A helper function to send confirmation email
 async function sendEmailConfirmation(user) {
   const newToken = jwt.sign({ id: user._id, type: 'email_confirmation' }, jwtSecret, {
     expiresIn: '3d',
@@ -31,6 +25,14 @@ async function sendEmailConfirmation(user) {
     { emailConfirmationToken: newToken, emailConfirmed: false }
   );
   return;
+}
+
+async function createUser(data) {
+  Object.assign(data, {
+    emailConfirmed: true,
+  });
+  const user = await User.create(data);
+  return user;
 }
 
 async function signup(data) {
@@ -59,7 +61,7 @@ async function emailConfirmation(token) {
   if (!user) {
     throw new NotFound('user not found');
   }
-  
+
   await User.findOneAndUpdate(
     { _id: id, emailConfirmationToken: token },
     { emailConfirmed: true, emailConfirmationToken: null }
@@ -100,6 +102,32 @@ async function login(data) {
   };
 }
 
+async function getUser(userId, isAdmin) {
+  if (!userId) {
+    throw new BadRequest('userId is required');
+  }
+  if (isAdmin) {
+    const user = await User.findById(userId);
+    if (!user) {
+      throw new NotFound('user not found');
+    }
+    return user;
+  }
+  const user = await User.findById(userId).select(['_id', 'name', 'email', 'role']);
+  if (!user) {
+    throw new NotFound('User not found');
+  }
+  return user;
+}
+
+async function getUsers({limit, skip}) {
+  const users = await User.find({}, null, {
+    limit,
+    skip,
+  });
+  return users;
+}
+
 async function updateUser(data, userId) {
   if (data._id) {
     delete data._id;
@@ -113,41 +141,41 @@ async function updateUser(data, userId) {
     await sendEmailConfirmation({ ...user, email: data.email });
   }
   await User.findByIdAndUpdate(userId, data);
-  return user;
+  return;
 }
 
 async function deleteAccount(userId) {
-  console.log('Starting acoount removal');
+  if (!userId) {
+    throw new BadRequest('user id is required');
+  }
+
   const user = await User.findById(userId);
   if (!user) throw new NotFound('user not found');
 
   const session = await startSession();
   try {
     session.startTransaction();
-    console.log('Starting removal of associated restaurants');
     const restaurants = await Restaurant.find({ owner: userId }, null, {
       session,
     });
     if (restaurants.length > 0) {
       await Restaurant.deleteMany({ owner: userId }, { session });
     }
-    console.log('Starting removal of associated reviews');
     const reviews = await Review.find({ reviewer: userId }, null, { session });
     if (reviews.length > 0) {
       await Review.deleteMany({ reviewer: userId }, { session });
     }
-    console.log('Removing account');
     await User.findByIdAndDelete(userId, { session });
     await session.commitTransaction();
     session.endSession();
     console.log('Succesfully removed the account');
     return;
   } catch (err) {
-    console.log('Transaction failed, reverting changes');
+    console.log('Removal failed, reverting changes');
     await session.abortTransaction();
     session.endSession();
     console.log(err.message);
-    throw new InternalServerError('Session failed');
+    throw new InternalServerError('Remove user session failed');
   }
 }
 
@@ -157,6 +185,8 @@ module.exports = {
   login,
   emailConfirmation,
   resendConfirmationEmail,
+  getUser,
+  getUsers,
   updateUser,
   deleteAccount,
 };
